+++ kornia/augmentation/_2d/intensity/__init__.py
@@ -15,7 +15,7 @@ from kornia.augmentation._2d.intensity.gaussian_noise import RandomGaussianNoise
 from kornia.augmentation._2d.intensity.grayscale import RandomGrayscale
 from kornia.augmentation._2d.intensity.hue import RandomHue
 from kornia.augmentation._2d.intensity.invert import RandomInvert
-from kornia.augmentation._2d.intensity.randomjpeg import RandomJPEG
+from kornia.augmentation._2d.intensity.jpeg import RandomJPEG
 from kornia.augmentation._2d.intensity.median_blur import RandomMedianBlur
 from kornia.augmentation._2d.intensity.motion_blur import RandomMotionBlur
 from kornia.augmentation._2d.intensity.normalize import Normalize
