+++ kornia/augmentation/_2d/intensity/jpeg.py
